Refactor blocks.blocks to blocks
