--- docs/source/foundry.core.blocks.rst
@@ -1,37 +1,0 @@
-foundry.core.blocks package
-===========================
-
-Submodules
-----------
-
-foundry.core.blocks.Block module
---------------------------------
-
-.. automodule:: foundry.core.blocks.Block
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-foundry.core.blocks.BlockGroup module
--------------------------------------
-
-.. automodule:: foundry.core.blocks.BlockGroup
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-foundry.core.blocks.util module
--------------------------------
-
-.. automodule:: foundry.core.blocks.util
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: foundry.core.blocks
-   :members:
-   :undoc-members:
-   :show-inheritance:
